Address comments from biweekly meeting (add enums.py file and update Axis enum docstring
