+++ modin/experimental/core/execution/native/implementations/omnisci_on_native/dataframe/dataframe.py
@@ -14,7 +14,7 @@
 """Module provides ``OmnisciOnNativeDataframe`` class implementing lazy frame."""
 
 from modin.core.dataframe.pandas.dataframe.dataframe import PandasDataframe
-from modin.core.dataframe.base.dataframe.dataframe import Axis, JoinType
+from modin.core.dataframe.base.dataframe.enums import Axis, JoinType
 from modin.experimental.core.storage_formats.omnisci.query_compiler import (
     DFAlgQueryCompiler,
 )
@@ -808,7 +808,7 @@ class OmnisciOnNativeDataframe(PandasDataframe):
         ----------
         other : OmnisciOnNativeDataframe
             A frame to join with.
-        how : str or modin.core.dataframe.base.dataframe.JoinType, default: JoinType.INNER
+        how : str or modin.core.dataframe.base.enums.JoinType, default: JoinType.INNER
             A type of join.
         left_on : list of str, optional
             A list of columns for the left frame to join on.
@@ -825,6 +825,7 @@ class OmnisciOnNativeDataframe(PandasDataframe):
         OmnisciOnNativeDataframe
             The new frame.
         """
+        how = JoinType(how)
         assert (
             left_on is not None and right_on is not None
         ), "Merge with unspecified 'left_on' or 'right_on' parameter is not supported in the engine"
@@ -1134,7 +1135,7 @@ class OmnisciOnNativeDataframe(PandasDataframe):
 
         Parameters
         ----------
-        axis : int or modin.core.dataframe.base.dataframe.Axis
+        axis : int or modin.core.dataframe.base.enums.Axis
             The axis to concatenate along.
         other_modin_frames : list of OmnisciOnNativeDataframe
             Frames to concat.
